requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
Flask==3.1.0
itsdangerous==2.2.0
pycryptodome==3.22.0
Blinker==1.9.0 
Click==8.1.8 
Colorama==0.4.6
Coverage==7.8.0